add comment delete/modify security

+++ routes/comment.js
@@ -25,17 +25,50 @@ router.post('/comment/add', function(req, res, next) {
 /* Get delete comment */
 router.get('/comment/delete', function(req, res, next) {
   if(req.urlParams.id) {
-    var deleteCommentQuery = "delete from comment " +
-      "where comment_id=" + req.urlParams.id + ";";
+    if (req.session.is_admin) {
+      var deleteCommentQuery = "delete from comment " +
+        "where comment_id=" + req.urlParams.id + ";";
 
-    res.locals.connection.query(deleteCommentQuery, function (error, results, fields) {
-      if(error != null) {
-        console.log(error);
-        res.render('articles', {'message': 'An error happened... Try again later!', 'is_admin': req.session.is_admin});
-      } else {
-        res.render('articles', {'message': 'Your comment is deleted!', 'is_admin': req.session.is_admin});
-      }
-    });
+      res.locals.connection.query(deleteCommentQuery, function (error, results, fields) {
+        if(error != null) {
+          console.log(error);
+          res.render('articles', {'message': 'An error happened... Try again later!', 'is_admin': req.session.is_admin});
+        } else {
+          res.render('articles', {'message': 'Your comment is deleted!', 'is_admin': req.session.is_admin});
+        }
+      });
+    } else {
+      //Verify it is its own comment
+      var selectCommentAuthorIdQuery = 'select author_id from comment ' +
+        'where comment_id=' + req.urlParams.id + ";";
+
+      res.locals.connection.query(selectCommentAuthorIdQuery, function (error, author, fields) {
+        if (error != null) {
+          console.log(error);
+          res.render('articles', {'message': 'An error happened... Try again later!', 'is_admin': req.session.is_admin});
+        }
+        if (author.length > 0) {
+          if (author[0].author_id === req.session.user_id) {
+            var deleteCommentQuery = "delete from comment " +
+              "where comment_id=" + req.urlParams.id + ";";
+
+            res.locals.connection.query(deleteCommentQuery, function (error, results, fields) {
+              if(error != null) {
+                console.log(error);
+                res.render('articles', {'message': 'An error happened... Try again later!', 'is_admin': req.session.is_admin});
+              } else {
+                res.render('articles', {'message': 'Your comment is deleted!', 'is_admin': req.session.is_admin});
+              }
+            });
+          } else {
+            res.render('articles', {'message': 'It is not your comment', 'is_admin': req.session.is_admin});
+
+          }
+        } else {
+          res.render('articles', {'message': '\'An error happened... Try again later!', 'is_admin': req.session.is_admin});
+        }
+      });
+    }
   } else {
     res.redirect('/articles');
   }
@@ -44,18 +77,37 @@ router.get('/comment/delete', function(req, res, next) {
 /* Get modify comment */
 router.get('/comment/modify', function(req, res, next) {
   if(req.urlParams.id) {
-    var selectCommentQuery = 'SELECT comment_id,content from comment ' +
-      "where comment_id=" + req.urlParams.id + ";";
+    //Verify it is its own comment
+    var selectCommentAuthorIdQuery = 'select author_id from comment ' +
+      'where comment_id=' + req.urlParams.id + ";";
 
-    res.locals.connection.query(selectCommentQuery, function (error, results, fields) {
-      if(error != null) {
+    res.locals.connection.query(selectCommentAuthorIdQuery, function (error, author, fields) {
+      if (error != null) {
         console.log(error);
         res.render('articles', {'message': 'An error happened... Try again later!', 'is_admin': req.session.is_admin});
       }
-      if(results.length > 0) {
-        res.render('comment', {'comment': results[0], 'is_admin': req.session.is_admin});
+      if (author.length > 0) {
+        if (author[0].author_id === req.session.user_id) {
+          var selectCommentQuery = 'SELECT comment_id,content from comment ' +
+            "where comment_id=" + req.urlParams.id + ";";
+
+          res.locals.connection.query(selectCommentQuery, function (error, results, fields) {
+            if(error != null) {
+              console.log(error);
+              res.render('articles', {'message': 'An error happened... Try again later!', 'is_admin': req.session.is_admin});
+            }
+            if(results.length > 0) {
+              res.render('comment', {'comment': results[0], 'is_admin': req.session.is_admin});
+            } else {
+              res.render('articles', {'message': 'This comment does not exists', 'is_admin': req.session.is_admin});
+            }
+          });
+        } else {
+          res.render('articles', {'message': 'It is not your comment', 'is_admin': req.session.is_admin});
+
+        }
       } else {
-        res.render('articles', {'message': 'This comment does not exists', 'is_admin': req.session.is_admin});
+        res.render('articles', {'message': '\'An error happened... Try again later!', 'is_admin': req.session.is_admin});
       }
     });
   } else {
@@ -66,17 +118,36 @@ router.get('/comment/modify', function(req, res, next) {
 /* Post modify comment */
 router.post('/comment/modify', function(req, res, next) {
   if(req.urlParams.id) {
-    var updateCommentQuery = "update comment " +
-      "set content ='" + req.body.content + "'," +
-      "date = NOW() " +
-      "where comment_id=" + req.urlParams.id + ";";
+    //Verify it is its own comment
+    var selectCommentAuthorIdQuery = 'select author_id from comment ' +
+      'where comment_id=' + req.urlParams.id + ";";
 
-    res.locals.connection.query(updateCommentQuery, function (error, results, fields) {
-      if(error != null) {
+    res.locals.connection.query(selectCommentAuthorIdQuery, function (error, author, fields) {
+      if (error != null) {
         console.log(error);
         res.render('articles', {'message': 'An error happened... Try again later!', 'is_admin': req.session.is_admin});
+      }
+      if (author.length > 0) {
+        if (author[0].author_id === req.session.user_id) {
+          var updateCommentQuery = "update comment " +
+            "set content ='" + req.body.content + "'," +
+            "date = NOW() " +
+            "where comment_id=" + req.urlParams.id + ";";
+
+          res.locals.connection.query(updateCommentQuery, function (error, results, fields) {
+            if (error != null) {
+              console.log(error);
+              res.render('articles', {'message': 'An error happened... Try again later!', 'is_admin': req.session.is_admin});
+            } else {
+              res.render('articles', {'message': 'The comment has been updated!', 'is_admin': req.session.is_admin});
+            }
+          });
+        } else {
+          res.render('articles', {'message': 'It is not your comment', 'is_admin': req.session.is_admin});
+
+        }
       } else {
-        res.render('articles', {'message': 'The comment has been updated!', 'is_admin': req.session.is_admin});
+        res.render('articles', {'message': '\'An error happened... Try again later!', 'is_admin': req.session.is_admin});
       }
     });
   } else {
